requirements.txt
Django>=4.2,<5.0
django-recurrence>=1.11.0
python-dateutil>=2.8.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
whitenoise>=6.6.0
